requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
ruptures>=1.1.9
scikit-learn>=1.3.0
lightgbm>=4.0.0
shap>=0.44.0
altair>=5.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0

